Some fixes

+++ source/mcu/imx8mn/src/stats.h
@@ -14,8 +14,8 @@ typedef struct {
 } AdcStats;
 
 typedef struct {
-    size_t buff_was_empty;
-    size_t buff_was_full;
+    uint32_t buff_was_empty;
+    uint32_t buff_was_full;
 } DacWfStats;
 
 typedef struct {
@@ -26,7 +26,7 @@ typedef struct {
     uint32_t max_intrs_per_sample;
     AdcStats adcs[SKIFIO_ADC_CHANNEL_COUNT];
     DacWfStats dac_wf;
-    size_t adc_buff_was_full[SKIFIO_ADC_CHANNEL_COUNT];
+    uint32_t adc_buff_was_full[SKIFIO_ADC_CHANNEL_COUNT];
 } Statistics;
